chore(opendal): Chore with change in opendal

+++ app/src/infrastructure/repository/storage/move_registration.rs
@@ -74,7 +74,7 @@ impl LeaseRepository<MoveRegistration> for RedisRepo {
         self.query(&Cmd::pset_ex(
             key,
             serde_json::to_string(&entity).unwrap(),
-            ttl as usize,
+            ttl as u64,
         ))
         .await?;
         Ok(())
@@ -90,7 +90,7 @@ impl LeaseRepository<MoveRegistration> for RedisRepo {
         self.query(&Cmd::pset_ex(
             format!("{user_id}_{key}"),
             serde_json::to_string_pretty(&entity)?,
-            ttl as usize,
+            ttl as u64,
         ))
         .await?;
         Ok(entity.id)
